refactor: consolidate molecule pair classes into single module
