+++ simba/core/training/train_utils.py
@@ -8,8 +8,7 @@ from rdkit import Chem
 from tqdm import tqdm
 
 from simba.core.chemistry.tanimoto import Tanimoto
-from simba.core.data.molecular_pairs import MolecularPairsSet
-from simba.core.data.molecule_pairs_opt import MoleculePairsOpt
+from simba.core.data.molecule_pairs import MolecularPairsSet, MoleculePairsOpt
 from simba.core.data.preprocessing import PreprocessingUtils
 from simba.core.data.spectrum import SpectrumExt
 from simba.utils.binning import round_to_ordinal
